requirements.txt
Pillow
pyyaml>=5.1
sentencepiece
tensorboard
tensorboardX
textgrid
pytest
flake8==3.8.2
flake8-bugbear
flake8-comprehensions
flake8-executable
flake8-pyi==20.5.0
mccabe
pycodestyle==2.6.0
pyflakes==2.2.0
torch>=2.2.2
torchaudio>=2.2.2
tqdm
deepspeed>=0.14.0
librosa
openai-whisper
pre-commit==3.5.0
langid
# rev
typeguard==2.*
torch_optimizer
numpy<2
gputil
wandb
boto3
causal-conv1d>=1.4.0
# mamba-ssm -> install from rev-mamba repository
# optimized_transducers -> install from rev-optimized_transducers